Increased performance of Fortran codes, added reading of new TSHS format
- The new transiesta code has a reformmated TSHS file which
  drastically reduces the disk usage.
  Remark:
   ONLY the FORTRAN IO will read the content correctly.
   The python equivalent IO will only read the "old" TSHS format.
   It should be quite easy to do the conversion to also enable
   reading of newer versions, but I think it would be better to
   require the users to ALWAYS use the Fortran routines. No?

- The conversion of data to Inelastica units are now performed in
  Fortran

- The data from the TSHS files are now saved in Fortran order
  which means that passing the arrays to the Fortran routines
  are better aligned in memory. For large systems this should
  increase performance.

- Expansion of the self-energies have an improved performance 
  when no repetition is used. 
  Using the Fortran routines increases the memory used when repetition
  is used, however the entire expansion is performed in Fortran
  which should speed up the execution.

- In the setup.py file I have added requirements for installed 
  versions.
  I am not fully sure about the versions that should be the required
  versions.
  Currently they are:
    numpy >= 1.6
    ScientificPython >= 2.6

- Added test of reading and comparing the old TSHS vs. the new TSHS
  file format.

+++ package/F90/surfaceGreen.f90
@@ -1,7 +1,7 @@
   ! Calculates the surface Green's function for the electrodes
   ! Handles both the left and right one
   subroutine surfaceGreen(no,ZE,H00,S00,H01,S01,accur,is_left,GS)
-       
+
 ! ***************** INPUT **********************************************
 ! integer     no      : Number of orbitals in the electrode
 ! complex(dp) ZE      : The energy of the Green's function evaluation
@@ -14,40 +14,40 @@
 ! **********************************************************************
     implicit none
 
-    complex*16, parameter :: z_1  = dcmplx(1._8,0._8)
-    complex*16, parameter :: z_m1 = dcmplx(-1._8,0._8)
-    complex*16, parameter :: z_0  = dcmplx(0._8,0._8)
+    complex(kind=8), parameter :: z_1  = dcmplx(1._8,0._8)
+    complex(kind=8), parameter :: z_m1 = dcmplx(-1._8,0._8)
+    complex(kind=8), parameter :: z_0  = dcmplx(0._8,0._8)
 
 ! ***********************
 ! * INPUT variables     *
 ! ***********************
     integer,     intent(in) :: no
-    complex*16, intent(in) :: ZE 
-    complex*16, intent(in) :: H00(no*no),S00(no*no)
-    complex*16, intent(in) :: H01(no*no),S01(no*no)
+    complex(kind=8), intent(in) :: ZE 
+    complex(kind=8), intent(in) :: H00(no*no),S00(no*no)
+    complex(kind=8), intent(in) :: H01(no*no),S01(no*no)
 
-    real*8   , intent(in) :: accur
+    real(kind=8)   , intent(in) :: accur
     logical,     intent(in) :: is_left
 
 ! ***********************
 ! * OUTPUT variables    *
 ! ***********************
-    complex*16, intent(out) :: GS(no*no)
+    complex(kind=8), intent(out) :: GS(no*no)
 
 ! ***********************
 ! * LOCAL variables     *
 ! ***********************
-    complex*16, allocatable, target :: zwork(:)
+    complex(kind=8), allocatable, target :: zwork(:)
     integer :: nom1, no2, nosq
     integer :: ierr             !error in inversion
     integer :: i,j,ic,ic2
     logical :: LEFT, as_first
 
-    real*8 :: ro
+    real(kind=8) :: ro
 
     ! on the stack...
     integer :: ipvt(no)
-    complex*16, dimension(:), pointer :: rh,rh1,w,alpha,beta,gb
+    complex(kind=8), dimension(:), pointer :: rh,rh1,w,alpha,beta,gb
 
     ! The inelastica way is opposite, hence we have to reverse it here.
     LEFT = .not. is_left
